remove invalid cast when getting default.
2006-02-17  Larry Ewing  <lewing@novell.com>

	* src/Preferences.cs: remove invalid cast when getting default.

+++ src/Preferences.cs
@@ -114,7 +114,7 @@ namespace FSpot
 			case ICON_VIEW_POSITION:
 				return null;
 			case SCREENSAVER_TAG:
-				return  (uint) 1;
+				return 1;
 			}
 
 			return null;
